+++ srcs/ft_mat_get.c
@@ -6,13 +6,13 @@
 /*   By: mmerabet <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/01 21:10:27 by mmerabet          #+#    #+#             */
-/*   Updated: 2018/02/07 17:04:57 by mmerabet         ###   ########.fr       */
+/*   Updated: 2018/02/07 21:20:51 by mmerabet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_math.h"
 
-float	*ft_mat_get(const t_mat *mat, size_t x, size_t y)
+double	*ft_mat_get(const t_mat *mat, size_t x, size_t y)
 {
 	if (!mat || !mat->matrix)
 		return ((void *)0);
